+++ src/main/java/com/github/marbor/shortcutsstats/StatsToolWindow.java
@@ -20,8 +20,6 @@ import javax.swing.JScrollPane;
 
 import static com.github.marbor.shortcutsstats.ShortcutsListener.UNKNOWN_SHORTCUT_DESCRIPTION;
 import static com.github.marbor.shortcutsstats.TextUtils.makeHugeNumberShorter;
-import static com.github.marbor.shortcutsstats.TextUtils.timeOrTimes;
-import static java.util.Comparator.comparingLong;
 import static java.util.Optional.ofNullable;
 
 public class StatsToolWindow implements Observer {
@@ -57,7 +55,6 @@ public class StatsToolWindow implements Observer {
 
         shortcutsStatistics.getShortcuts()
                 .stream()
-                .sorted(comparingLong(Shortcut::count).reversed())
                 .map(s -> new ShortcutView(getDisplayText(s), getDescription(s)))
                 .forEach(model::addElement);
 
@@ -85,7 +82,7 @@ public class StatsToolWindow implements Observer {
     }
 
     private String getDisplayText(Shortcut shortcut) {
-        return shortcut.shortcut() + " pressed " + shortcut.count() + " " + timeOrTimes(shortcut.count());
+        return String.format("<html>%d - <b>%s</b> (%s)</html>", shortcut.count(), shortcut.shortcut(), shortcut.description());
     }
 
     private String getDescription(Shortcut shortcut) {